move `removeVariableDeclarator` to a new file
